+++ BlogManagement.Model/Model/Sys/T_Sys_UserGroup.cs
@@ -18,17 +18,17 @@ namespace BlogManagement.Model
         ///描述：标识
         ///</summary>
         [Column(IsIdentity = true, IsPrimary = true)]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         ///<summary>
         ///描述：用户标识
         ///</summary>
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         ///<summary>
         ///描述：组标识
         ///</summary>
-        public int Groupid { get; set; }
+        public long Groupid { get; set; }
 
         ///<summary>
         ///描述：状态 - 0：禁用；1：正常
@@ -38,7 +38,7 @@ namespace BlogManagement.Model
         ///<summary>
         ///描述：创建人
         ///</summary>
-        public int Creater { get; set; }
+        public long Creater { get; set; }
 
         ///<summary>
         ///描述：创建时间
@@ -48,7 +48,7 @@ namespace BlogManagement.Model
         ///<summary>
         ///描述：修改人
         ///</summary>
-        public int Updater { get; set; }
+        public long Updater { get; set; }
 
         ///<summary>
         ///描述：修改时间
